Fix some minor problems with pull request #1.
TranslationUnit._is_within() has been changed to an ordinary method,
it seemed to be slightly preferable to adding the @staticmethod decorator.

Replaced system.h with time.h in the test data, system.h is not a standard C
header and relying on its presence can therefore result in an unnecessary
test failure.